refactor: ACI-5128 co-locate single-consumer fixes; promote shared ones
- fix_self_exec.go split:
  - daemonUpFix + defaultDaemonUp → fix_daemon_up.go (shared: ccache + xcelerate)
  - updateFix + runSelf + defaultRunSelf → check_cli_version.go (sole consumer)
  - daemonRestartFix + defaultDaemonRestart → check_xcelerate.go (sole consumer)
- activateWizardFix + defaultLaunchActivateWizard moved out of
  check_auth_backend.go into fix_activate_wizard.go (shared: auth + auth-backend)
